+++ packages/widgetbook_generator/lib/metadata.dart
@@ -1,2 +1,2 @@
 /// The version as in pubspec.yaml
-const packageVersion = '3.5.0';
+const packageVersion = '3.6.0';