+++ packages/widgetbook_generator/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- **EXPERIMENTAL**: Use params' default values for `StoryArgs.fixed` constructor. ([#1074](https://github.com/widgetbook/widgetbook/pull/1074))
 - **EXPERIMENTAL**: Support `EnumArg`. ([#1073](https://github.com/widgetbook/widgetbook/pull/1073))
 
 ## 3.5.0