+++ packages/widgetbook_generator/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **EXPERIMENTAL**: Use params' default values for `StoryArgs.fixed` constructor. ([#1074](https://github.com/widgetbook/widgetbook/pull/1074))
 - **EXPERIMENTAL**: Support `EnumArg`. ([#1073](https://github.com/widgetbook/widgetbook/pull/1073))
+- **EXPERIMENTAL**: Support positional `Arg.value` parameter. ([#1077](https://github.com/widgetbook/widgetbook/pull/1077))
 
 ## 3.5.0